Lem catches unexpected "Failures" but I want stack traces!
On a side note, I think we should integrate our hack into mainline Lem. The
mechanism our hack is using already exists for ppcmem so we are not really
introducing any additional ugliness. On the plus side, we could use (and test!)
an up-to-date Lem version, provide useful feedback to our developers, etc.