Streamlining of definitions, round 2.
We now have a separate predicate for checking whether a an expression e2 can
be assign to an lvlue of type ty1. Namely, checkAssignable P G S ty1 e2. It is
reused to type-check the arguments of a call expressions. The benefits are:
(1) it's closer to the standard and (2) we no longer need to introduce a fresh
variable in the typing rule for function arguments.

The next change is similar in spirit. I have separate the checking from the
synthesis of types in binary arithmetic expressions. They're reused for
compound assignments. Again, this is much closer to the standard text.

Lastly, I have proved I strengthening/weakening lemma (for
expressions). Hopefully I won't need it but it was a good test case to see
whether my changes made proofs any easier (conclusion: yes, substantially).