Make eval be guaranteed to terminate.
There were previously three cases in eval (the function that evaluates pure Core expressions) where eval was called recursively on an value that was not a subterm of its argument: when evaluating Elet, Ecall and Ewseq.  This meant that it was possible for eval not to terminate, which was a problem for the Coq backend.

This commit removes those three recursive calls.  Instead, eval now returns the partially-evaluated expression, and relies on the caller re-invoking eval (potentially multiple times) to finish the evaluation.  Essentially eval has changed from being a big-step evalution, to being a "medium-step" evaluation.

The remaining changes in this commit are to update all the callers of eval (namely the red2_xxx functions) correspondingly.  Because we require those functions to be terminating as well, they cannot simply call eval in a loop.  Instead, the pattern is (in pseudo-code):

  red2_foo (Efoo pure_expr) =
    if is_value pure_expr then
       return (result of computation using the value)
    else
       return (Efoo (eval pure_expr)) # return the argument with the pure expression replaced by the result of (partially) evaluating it