Make pretty-printing of core programs prettier.  In particular, we now get output like:
  let a = 0 in
  let b = 1 in
  let c =
    let d = 2 in
    let e = 3 in 4 end end in
  let f = let g = 5 in 6 end in
  let i =
    let j =
      let k = 7 in 8 end in 9 end in
    10 end in
  expression

where the indentation level reflects the nesting of the first sub-expression of 'let a = e1 in e2 end' expressions.

Line-breaks can still only occur due to let..end or if..then..else.  To be prettier still, we should allow line breaks in lists/tuples, with appropriate additions to the indentation level.