[BMC] Support OOB-by-one ptrs
TODO: for consistency with Core semantics, PtrValidForDeref should not
be doing a range check. This should be done at the time of load/stores.
We need to modify the memory model bindings to deal with what happens
when we read from an OOB ptr (such that the binding is still satisfiable
and the error is only at the point of checking VCs to go along with the
BMC pipeline; otherwise we can't extract all executions).