Generalise exact match for qpoint missing fields
Handle a specific case when a qpoint in the resource context
is missing a particular field, e.g. an array arr which has
had arr[i] adjusted and is now missing it. After the exact
match phase, the remaining "needed" part will contain the
conjunct "ivar = i" for the index var ivar. This gives us the
hint that we should now try exact matches of arr[i].