Keep values/sym_eqs equalities in context
Constraints of the form "x = ..." where x is a symbol/variable
are now kept in a map in a separate field of the context.

This is similar to what the simplifier already did when it
collated such equality constraints into its "values" variable
before simplification. The key advantages are (1) this is done
in the context management rather than each time the simplifier
starts up and (2) these equalities are not shown to Z3. This
seems to have a substantial impact on performance.