First steps towards a nested timing variant
Yet another variation of timing, given that we have a slow
problem we're trying to understand. This tries to build a
hierarchical tree of timing info, with nested start/end
markers, which might be useful for including some contextual
info that exists higher in the hierarchy (that is, the
function call stack). For the moment, it's not working
particularly well.