Tweak nested time log to ~JSON
Maybe it would make sense if the log parsed as a native
JSON object and/or python value? Some progress in that
direction, but it isn't exactly working out great.

(It's annoying that JSON doesn't seem to permit a trailing
comma, e.g. [ 1, 2, 3, ], so that the printer needs either
state or silly hacks to avoid printing too many commas.)