Explain missing resource, clauses in state.html
When the type error is the a resource was missing, add a line
to state.html formatting the missing resouce in the same style.

Also, attempt to locate any logical predicate clauses that can
be used to pack the missing resource (if it is a logical
predicate) or unpack objects including a resource of the same
type. The filtering basically works, but the pretty-printing
of these is a mess.