Add qpoint->point request recursion
Recognise some cases where a qpoint request necessarily
requires some point elements, e.g. if one of the elements
of an array had been claimed and become a separate resource.
In these cases, recurse to performing a full point_request
call.

This makes a difference because point_request will then
recurse to fold_struct or fold_array if that might allow it
to succeed. Concretely, this means that explicit fold_struct
and fold_array statements are needed less often.