Add code for interpreting resource spans
This is now added to the rendering of resource values in
(HTML) state reports. It's also easy to add a gadget that
spots problematic intersecting spans. It's then clear what
to do if these are Point resources/requests. However, this
should really handle QPoint resources/requests as well,
which requires guessing the index to act on, which is a lot
more tricky.