Add comparison feature to CN time log parser
The python in scan_cn_json.py can now do a comparison of timing
between a pair of time logs produced by CN runs. This handles
incomplete files, so comparison of timing is possible before
CN completes or after an error.

The code tries to name and highlight interesting discrepancies,
which for the moment is a bit crude.