Handle impossible scenarios at error terms
CN previously assumed that the logical state was always satisfiable,
and tried to maintain this by checking for impossible/unsat on any
if-branch or case-split or similar. The code for checking various errors
then assumed that this is a concrete error to be reported rather than
an impossible scenario.

However, unsatisfiable logical states can also arise when information
is added after a function call or an unpack. It's better to handle these
cases by returning void.