Add lemma-gen mode and merge calling options
Adjust the new-elaboration/call-by-value changes to conversions.ml
to be switchable, so the retype/conversions mechanism can parse
assertions as appropriate to the old (call-by-reference) function
calling style or the new (call-by-value) style. This is important
as support for the new style is incomplete and is expected to stay
that way until changes to subfunction/label calling are completed.

Also, the very beginnings of some lemma-generation code, including
a toplevel switch to generate lemma files rather than check type
signatures.