First working lemmata sweep
It's working in the sense of "not crashing" as opposed to
"doing anything useful". Still, for the first time, the cbv
variant of the page_alloc test is processed (by adding a
hacky flag to drop label annotations) and the pure subset
of the lemma types are copied into something approximating
Coq syntax.