Setup for tracing mu steps in errors
The idea is that a trace document will be produced in the same
style as the current state.html mechanism, and at the same times.
This requires storing the trace of mu-step context changes (in
the typing monad), passing this to some type error functions, and
moving the mucore instantiation out of Retype and into its own file
to avoid a Retype->TypeError->Trace dependency loop.

The sheer number of typeError message constructors is a bit of a
hurdle here, so the trace is only passed in a representative instance.

Still to do, actually format the trace.