CN: broader notion of 'constant' in NIA check
CN checks that specifications do not encode non-linear integer
arithmetic, e.g. the right-hand-side of any division should be
an integer constant. However, the existing syntactic check that
the rhs is an explicit numeral is problematic for the goal of
making more use of C macro definitions etc. Instead, check that
no free variables or uninterpreted logical predicates are
present, which ought to imply the value is constant, and the
simplifier probably makes that explicit.