Add minimal support for C->logical coercion
As before, the idea is to use an (irrelevant) C function to
define a logical computation, and convert that to an index-term
body by doing some minimal symbolic evaluation. This is now
supported for the very most minimal of cases, where the body
is literally a return of a constant.