Permit logical predicates in iterated steps
This is similar to the recent 'more permissive NIA' change. The
idea is to be able to define a value, e.g. page_size (), by extracting
it from a C function, and then use that constant in the definition
of an iterated resource, e.g. the buddy page collections which have
page_size () (that is, 4096) as an offset.

For various resource-inference things to work, these expressions must
actually resolve to concrete constants we can evaluate.

Tricky part, the simplifier can do the expression simplification, but
isn't available during the construction of resources for dependency
and contextual reasons.

Instead, permit general index terms as the step of resource elements,
insist in the wellTyped phase that these be morally constants (in the
same sense as the previous NIA change, no free vars or undefined vals),
and then check (in resourceInference and in typing's resource-adds)
that iterated resources we are operating on actually have constant
steps.