CN: parsing of datatypes working now
There is some potential confusion in whether the id->symbol conversion,
the resolution of strings to unique identifiers, is done during the
Cerberus desugaring (cabs -> ail) phase or later within the CN backend
(in compilePredicates). The final answer for datatypes is that datatype
names and constructors are resolved in cabs->ail, but that datatype
member names are resolved later in compilePredicates.