CN: datatypes in assertion syntax, retype global
Adds the datatype mechanisms to the assertion parser syntax,
approximately duplicating the changes made to the file parser
before.

This means the conversion/retype mechanism needs to be passed
the datatype info as well as predicate defs. This was getting
unwieldy, so some refactoring is done to create a 'global'
object earlier and pass it in to the retype phase.