CN: span updates, setup for predicate fold/unfold
The span code is updated, comes with a document, and is able to
detect a case where a requested resource predicate is null
(i.e. is definititely in a specific clause/case that claims no
resources). Rearrangement of check/resouceInference to allow the
span process to recurse into a spine-type action needed to fold
a predicate is underway.