CN: trace non-provability of resource pointer eqs
In addition to tracing unproven constraints and their evaluation
in the SMT solver model at debug level 6, also trace unproven
equalities of pointers in the resource inference loop at debug
level 9. This will create some noise when there are many resources
of the one type, but may help diagnose failures when it is clear
to an advanced user which is the correct candidate (e.g. when there
is only one candidate).