CN: deactivate 'subtract' step in spans
The 'subtract' step in the span logic was meant to help guarantee
termination. Unfortunately it doesn't work, as currently implemented.
The presence of an overlapping resource to be 'subtracted' from the
request might be a coincidence in the current model. We have though of
various possible fixes, but none are ideal, and the mechanism exists
to defend against a non-termination scenario that has never happened
yet. Meanwhile, it is creating a non-determinism scenario in which
unlikely problematic models lead to diverging behaviour.

This is just a temporary fix, a better version should hopefully follow.