CN: begin refactoring bind and pattern-match
Some adjustments to the code for bind_logical, bind_logically,
and pattern_match, to slightly reduce the number of logical
variables created, and to create them in code order.

The previous code was out of order as, when checking "let p <- x in y"
where p is a pattern, it first executed on "p", then "x", then "y".
This meant the expected return type of "x" was known before checking "x",
but it also meant variables were created out of order and then attached
to their "inputs" later by adding equalities. The new code adds a less
effectful computation of the expected input type of a pattern, allowing
a change of ordering.