Follow-up: less variables in pattern matching
It's now possible to simplify the pattern-match routine. Instead
of creating many variables and returning info about how to connect
them, it is given a pattern and value known to match that pattern,
and it binds variables to the actual results as needed.

In particular, add a gadget that names a value as a variable if it
is compound, but avoids creating new names for things that were
already symbols.