CN: name exprs returned in type-checking
When type checking "let p <- x in y", try to name the expression
that "x" returns when it is returned, rather than when it is
expanded against the pattern "p". This means its info can include
the mu-expression that gave rise to it, rather than just the
pattern it is bound at.