CN: first steps to model-based path pointer eqs
After a few rounds of discussion with the CN team, there is a plan to possibly
replace the 'span overlap' model-based logic with a 'equality to path-pointer'
mechanism. This would use models to suggest situations where a requested
pointer equals a pointer with a structural path e.g. p == &(q->array[i].x) .

This is a first implementation of the code that produces this suggestion.