CN: Z3 datatype mapping: restore some code
A Z3 constructor is a value in the same type (Z3.Func_Decl.func_decl) as any
other function symbol, and its definition consumes a Z3 symbol. However,
although it looks the same, it is a mistake to think that it *is* the same as
the function declaration that would be constructed from this symbol. Oops. When
encoding datatype ops, go back to fetching the correct func_decl in a
roundabout way rather than trying to directly make it. This fixes a strange bug.