CN: set up for pattern-match syntax
The syntax will look something like "x ?? Node {left = Node {}}"
and compile to a boolean which tests whether x has constructor
Node and x.left also has constructor Node. This can then be used
in if-then-else expressions "(x ?? Node {}) ? ... : ..." to emulate
a pattern match in a functional language.

The setup here is done, but no actual syntax is added to the parser.