CN: compile datatype member access in lemmata
The datatype member access T.left, which is standard in Z3's
datatype support and in CN's style of datatype support, can't
exist by default in Coq without evidence that T is a constructor
variant that has a T.left.

The workaround is to require the user to use the recognizing
pattern match from the previous patches to build an if-then-else
tree which is then compiled to a pattern match on the way into Coq.
TODO: better error messages when this goes wrong.