CN: nth_list/array_to_list as builtin
This tries to automate some of the array_to_list setup done
in tests/cn/tree16/as_mutual_dt by making array_to_list and
nth_list builtins, and providing the theorem that links them
automatically via a hook in the solver translation.

There is a cost: this re-purposes the list type and the existing
nthList operation for this purpose. The list type now maps to
an uninterpreted sort in Z3, rather than an accurate sort with no
supported operations.