CN: use auto-monomorphisation in an example
This as_auto_mutual_dt version of the tree16 example makes use
of the new feature and reduces the amount of lemma application
required down to just proving that the lookup functions can
exist as well-defined recursive functions. This version is
completed in Coq by using a library that is incomplete and
provided in a hacky way .. to revisit that.