CN: maintenance of the "slow solver" mechanism
Adjust some bits that had bit-rotted and/or never worked.

Slow passes of the incremental solver can now be saved. This will
usually never happen, since the timeout of the incremental solver
is low, but this allows the implementation to save all problems if
the slow threshold is set zero or negative.