CN: support for finite split of fun pointers
Adds support for splitting into cases when a function pointer is,
for instance, known to be equal to one of an enumerated list of
targets, e.g. (fp == fun1 || fp == fun2).

Also adds a builting function which expresses this,
(in_loc_list (fp, [fun1; fun2])).

Also adds a test case.