CN: history check for unseq
The official Core semantics for the unseq constructor requires race-detection,
e.g. when checking (unseq (x, y, z)), we need to raise errors if x and z both
claim the same resource in a race.

This is now detected by adding some history information to resources, so that
the write-sets of x, y, and z can be checked for intersection.

The problem remains that reads are treated like a sequence of
destroy/recreate-identical, which leads to errors about races that
aren't actually problematic.