CN: add simple path for unfold-struct
Adjust the resource inference path for unfold_struct to make use of
the map_and_fold_resources Unfolded case (previously unused) and
ensure that it preserves the history of the split-out resources.

Probably working, but doesn't solve the interesting test-case, since
that involves an unpack of a predicate as well as a struct.