CN: additional unpack needed for unseq
Fix the final failing test by adding an explicit unpack.

The interesting failing line is:
   b_tree_keys_node_lemma(p->even, p->odd);

The problem is, if the ownership of the struct is inside a
resource predicate, the type-checking of `p->even` triggers
an unpack of that predicate. Each unpack clause of a predicate
goes via spine judgements, and it was too complicated to make
the resulting resources not look "new". This leads to a race on
that resource, "created" at `p->even` and read at `p->odd`.