CN: fix a bug with id allocation
Adjusting map_and_fold_resources to be effectful (to permit
resource history writes) was a mistake. This leads to concurrent
confusion as to the "correct" value of the next-resource-index
value and is silly. Instead, add the history to the accumulator,
provide some specialised ops in context.ml, and fix the bug.

(This will hide some spurious races caused by dual allocation of
timestamps in a silly way.)