Magic tokens: toplevel CN decls in comments
It now (seems) to be working to wrap, for instance, a CN predicate
definition with /*@ @*/ at the top level.

The implementation is a giant hack. It reruns part of the parser
back over toplevel magic comments in the parse output. TODO: transition
to this format, deprecate the previous, and come up with a less hacky
implementation.