New CN resource inference.
Resources are eagerly auto-unpacked and packed on-demand. Iterated
separating conjunctions require some user input: users specify the
relevant indices and the resource inference freely moves ownership of
these indices in (on demand) and out (eagerly) of ISCs. The keyword is
currently called 'extract' but has to be renamed.

It's currently slow: ~60s vs ~25s on buddy allocator, to be optimised
based on footprint disjointness information. On this example it saves
19 lines of in-function annotation.