CN: restore some printing aid for SMT problems
An alternative print format gets rid of unreadable SMT
constant names (with spaces in them) that seem to confuse
some of the Z3 output. It still seems to be broken though,
with Z3's diagnostic SMT log containing operations on
datatype constructors that it complains about when given
as input syntax.