CN: make extern global symbols appear
In exactly the same was as function symbols and global constant
symbols lead to addresses that can be used in expressions, extern
symbols (of functions or globals) should be useable as well.

This works well enough for extern functions. If we start fiddling
with extern globals, we might need the alignment constraints to
be set up correctly.