CN: add an "inline" CN-statement.
It seems that, in various cases, it would be nice to have the type-checking at
a function call site simply include the semantics of the called function rather
than have to write out a specification for it that captures its semantics with
enough precision.

One use case (sketched here in tests/cn/cn_inline.c) is where a static inline
function is used in the computation of a constant. We don't want to have to
write out a specification of the function that explains exactly the arithmetic
it does, we just want the checking at the call site to reduce the computation
down to the constant.

The current version is designed to work by annotating the call site with this
information rather than annotating the function to be inlined at every call
site. This might be revisited.

Adjusting the parser to include the new statement required purging and
rebuilding the parser error-messages file. Apologies if anyone has manually
edited that file, if so, restore from history prior to this point.