CN: CLogicalFuns: support yet more syntax
Continue expanding the collection of Mucore syntax forms that CLogicalFuns
can map to CN syntax. It should probably be explained that this work began
with bd167f87a (two commits ago) which was pushed with no commit message by
mistake, whose main change was to refactor the mucore->index-term conversion
to handle if-then-else splits.