CN: handle some GCC builtins
The test for GCC's count-trailing-zeroes (CTZ) now passes. Some silly things
remain, for instance, in hindsight, handling CTZ and FFS (unary ops) in the
binop type was an odd choice. Inlining procs at the mucore level is also
probably quite unsafe, and should check that no "save" operations are being
introduced. This may become irrelevant if a pending change to the type/kind of
these builtin ops is introduced.