Add type_synonym syntax, retire uncommented syntax
With the change to datatype-matching, it is no longer convenient to use a
datatype as a type synonym for a named record. Instead, introduce an explicit
type_synonym keyword, which for now just does syntax substitution. Use this to
fix the syntax of the failing tree16 examples. The auto-version is still
failing for semantic reasons, which we might guess are to do with scoping of
datatype-match variables.

This lead to some confusion in the parser with the legacy enter_cn and exit_cn
lexer callbacks, which have now been removed. That means that the deprecated
toplevel syntax in which CN toplevel declarations like "predicate ..." appeared
without a wrapping comment is now retired entirely.