CN: move some problem-syntax walks to index-terms
These previously worked by walking over Z3 expressions, which is in some ways
problematic and currently broken by the datatype changes. This change moves the
problem-syntax walk to be done purely on index terms, which is generally
preferable. Unfortunately the failing example is still broken for essentially
the same reason, that the changes have placed it in a non-global context which
requires more care about binders.