CN: overcomplicated solution to local terms
Adjust the fold mechanisms from IndexTerms to track the pattern bindings that
give rise to the local symbols rather than just their names and types. This
makes it possible, in situations that are not under an actual quantifier, to
reconstruct a globally valid term by rebuilding a pattern match around it.

This fixes the auto-tree16 example.