CN: more normalisation via double-run of Welltyped
As resource predicates, logical functions and datatypes can recurse (without a
prototype/definition mechanism as in C), the check code pre-registers
simplified versions of these elements (essentially their signature) before
checking the bodies. Minor fix, the Welltyped code should be run on the
simplified version as well, so that, for instance, record field normalisation
has been done in the types.