CN: permit typedef tokens in CN annotations
Like enum consts, typedefs are a "natural" way of encoding a complex C type
(which we probably don't want to parse) into an atomic type name which it is
convenient to share in CN annotations. This will be more relevant when CN
distinguishes (in CN syntax) between different-width C integer types.