Rewrite bitvector->SMT conversion without wrapper
Replace the existing wrapping datatype, where signed bitvectors were kept in
a different (isomorphic) type to unsigned. This seems to contribute to poor
performance. Instead, signed and unsigned bitvectors (in CN) are mapped to the
one bitvector type (in Z3) with a tweak added to the reconstruction to insert
a cast where necessary.