CN: bv-solver: probably final investigations
This branch has explored the idea of doing bounds-inference on integer
sub-expressions of a logical problem, purely in the SMT solver mapping, for the
purposes of representing an integer problem using bitvectors and eventually
supporting bitvector ops. This will probably be the final commit. The problem,
as should perhaps have been obvious from the beginning, is that values in the
SMT problem include arithmetic values and logical values, but also logical
compound values (tuples and datatypes) that include integers, and this creates
numerous cases where bounds inference is nontrivial or coercion needs to be
pushed within values.