CN: bugfix to fold_with_adj: expand lets
Also includes a slightly faster implementation of let-unfolding.

There was a silly issue with the combined adjust/fold passing over some syntax
exponentially many times (for every combination of folded/unfolded of all the
let bindings it might be under), which is silly. Explicitly unfolding lets in
the adjust/fold implementation gets the buddy allocator back to normal runtime.