CN: add specs for some stdlib builtins
This fixes the builtin ctz/ffs example.

This works, for now. It would be good to also type-check the stdlib body
against the spec that is hand-written here, and it also be good to be able to
parse this stdlib spec rather than type out OCaml code to construct its packing
function.