CN: cLogicalFuns: try to handle no-peval case
Without partial evaluation, there are more kinds of terms present in the mucore
syntax, including terms we just want to ignore for these purposes. Begin
supporting more (e.g. function-pointer signature construction, which will get
ignored later) and simplifying out some arithmetic.

Still fails on a relevant test-case. The problem is that shift-left forms a
problematic pattern (probably also for final type-checking) resulting in
mod(x * (2 ^ n), max_int(ctype) + 1). The existing stuff will assign a type to
the RHS which is equivalent to ctype, so max_int(ctype) + 1 = 0, and the
modulus is nonsensical, and we're stuck.