CN: clean up check_conv_int and in_range confusion
Edit some code that was confused across basetypes.

Firstly, implement "is representable" checks as checks between constant numbers
and index-terms, rather than index-term/index-term. This means there's no need
to find an appropriate containing type. Either the upper bound for x is itself
representable in x's type, in which case it's a simple inequality, or it isn't,
in which case the inequality reduces to true or false. Ditto the lower bound.

I (TS) am convinced that the success case of conv_int, and wrapI, is equivalent
to just doing a bitvector cast.