Fix issues with if-1-0 in c->logical conversion
It turns out that there are at least two places where the 'if x then 1 else 0'
pattern appears, and they need to be treated with care. Some are used
internally and should be treated as signed int. Others are stored into values
of type bool/_Bool and should be treated as its storage type. Oops.