Return to much faster solver.
Previously, after the switch to bitvectors, the incremental solver got stuck on some SMT problems and we had to switch to a non-incremental solver to get these tests running, which is much slower on average. This is switching back to the incremental solver, but now with some simplifying pre-processing added.