Clean up various CN hacks in lexer/parser
This cleans up a lot of the lexer/parser mechanism for CN `/*@ @*/` magic
comments to use the one tricky mechanism. The initial tokeniser gathers a
collection of strings (plus their original location) as the magic comment
contents, and builds a custom lex-buffer from them for the CN re-parse. Tokens
quoted in error messages are reconstructed from the string collection rather
than the lex-buffer, since the latter is difficult or impossible.

This clears up a collection of historical attempts at fudging CN token
locations and pre-parsing (e.g. splitting on '@' and ';' characters prior to
lexing/parsing).

This also fixes the quoting of tokens in CN syntax errors, which was
consistently broken up until now, despite the many attempts at fudging. This
was the point of the exercise, although in hindsight, the actual problem could
have been fixed without the big adjustment.