CN: follow-up on function pointers
Retire a little more code in check.ml, and link to a feature in the typing
monad for finding a semantic equality. The equality to a function pointer in
the context now just needs to be provable, not an explicit syntactic equality,
which might make life a little easier. In practice this means some of the tests
require one less case split.