CN: add a cast to the ptr-diff check code
There was a TODO left in check.ml documenting confusion between the intptr_t
type (which is the address part of the pointer) and the ptrdiff_t type (which
is the signed type that encodes pointer differences). Adding a cast prevents
this just failing immediately (c.f. the added test). I'm fairly sure it doesn't
matter whether the subtraction is done at the intptr_t type and cast to
ptrdiff_t or whether the cast is done first. I've left the TODO, though, it's
still not clear this always results in the expected value, and worth thinking
about whether there should be a failure case.