CN: bugfix basetype checks for shifts/binops
The C shift op permits mismatching integer types. Some of the check code had
been adapted for this, but it looks like an incorrect check remained in the
signed (checking) case. Also, on re-reading, the check that the argument types
must be the same for all non-shift ops had been lost. Fixed.