CN: make resource unfolds wait lazily for query
Instead of checking for possible resource unfolds whenever a new resource or
logical constraint is added to the world, wait for the next query of the
resource state or the provability of anything. This reduces the number of
consistency checks that get done at incremental points.

This doesn't seem to break anything and it substantially speeds up the startup
time of CN on problems like the buddy allocator.