Add location info to terms
This is a best effort commit, the goal was to have the OCaml code simply
type-check and some location inforation attached to terms and availabe
for printing, debugging and other tooling. There are many choices to
revisit and likely many "errors" - it's actually not even clear what the
criteria should be for location information:

Option 1: Locations never lie - source locations always correspond to
something the user wrote or Cerberus elaborated. Whilst a noble goal,
this becomes potentially useless with the slightest rewrites.

Option 2: Locations track derivations - a location refers to the source
it was derived from or related to in some way. Is it possible to have
multiple origins?