Refactor the substitution API for easy renaming
Alpha-renaming is done frequently, and aside from being clunky, the
previous API also made it basically impossible to preserve the location
information for alpha-renamed variables (without resorting to very ugly
and brittle hacks).

This commit adds an extra possibility to the "payload" of a varible in
a substitution - previously it was just (Sym.t * IT.term), but now it's
(Sym.t  * [ `Rename of Sym.t | `Term of IT.term ]). This means the
location (and basetype) can be guaranteed to be preserved during
alpha-renaming (yay!) and is less to type at call sites.