CN: Unify handling and order of CN arguments
This was meant to help with moving the check for redeclared datatype
variables earlier (into elaboration, since it is purely syntactic)
and simplify later-stage code.

However, the CN backend seems to requires that such variables aren't
simply unique to the constructor, but to the entire datatype
declaration.

This is weird, and is probably an arbitrary restriction that should be
lifted, but it will require effort I cannot currently expend, hence this
is explained in this commit and in a comment.