Simplify CN frontend and reduce the number of constraints CN has to work with.
CN used to implicitly add `good<...>(...)` constraints for function
arguments, return values, Owned resource outputs in function
specifications to ensure values are within the representable range of
their C-types.

This used to be necessary due to the mismatch between the types of
integers in CN (unbounded mathematical integers) and C (fixed-width
integers).

This commit removes the default `good` constraints, which are no
longer required since the move to bitvectors. The remaining
implicitly-added constraints are "good-pointer" constraints to require
pointer alignment for Owned pointers.

Since disabling the VIP memory model was previously done using the
same mechanism, this commit changes how that is done (instead using a
unit-mapping for allocation IDs).