CN: Use flags to control mucore pp location output
The first is using the existing the -d<level> flag to output source
location information in the mucore pp ast file for level > 1.

The second is the flag --no-inherit-loc to see which source location is
provided directly by Core (and mark all the ones we infer from the
parent node as unknown).