CN: Refactor Astmt/Aexpr annot
It wasn't clear what the Astmt and Aexpr were used for (CN, to mark the
boundary of Ail statement/expression elaboration) and that at most one
of these can/should occur in a given annotation.

This should hopefully make subsequent commits for pretty-printing mucore
location information clearer and the pretty-printing output less noisy.