CN: Add minimal version of interactive symbolic stepping to error reports.
There's various things that need to be improved, most importantly,
showing the full function source code (with current position
highlighted) alongside the rest of the report, and fixing the
granularity of the steps: some steps are small steps in the underlying
Core program that have no visible effect in the information displayed
in the report, so clicking 'next' the website may seem unresponsive.

Note: at least on my machine, for some reason, Safari does not seem to
agree with the page anchors this uses.