CN: Fix #322 suggest alternative to pointer ==
As per https://github.com/rems-project/cerberus/issues/322
under the VIP memory model `==` in C is not the same as `==` in the spec
language (for pointers) and so there's a warning. This commit makes the
warning actionable with a concrete alternative sytnax to use which
makes explicit which type of equality is being used.