CN: Split out some well-formedness spec checking
This is so that most of the specs can be checked for well-formedness
before executable spec generation is run. It has the downside of
exposing the internal types of `checked` and `lemmata` but that may be
addressed later.

Note: the commit does not split out checking for loop invariants and CN
statements.